MESOS-2364 - Docker containerizer should pull every time if image tag is ":latest"
The "latest" tag for a docker image denominates the newest or most
recent version of a docker image. So the docker containerizer should
pull every time from the registry if the "latest" tag is encountered
to ensure that the most recent version is used.

Review: https://reviews.apache.org/r/31142